[update] ちょっと修正＆ 簡略化
