--- Assets/Project/Scripts/PlayScreen/TouchEvent.cs
@@ -1,39 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class TouchEvent : MonoBehaviour
-{
-    // Start is called before the first frame update
-    void Start()
-    {
-        
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        
-    }
-
-        public void TouchCheckD()
-    {
-        Debug.Log("lane D touched");
-    }
-    public void TouchCheckF()
-    {
-        Debug.Log("lane F touched");
-    }
-    public void TouchCheckG()
-    {
-        Debug.Log("lane G touched");
-    }
-    public void TouchCheckH()
-    {
-        Debug.Log("lane H touched");
-    }
-    public void TouchCheckJ()
-    {
-        Debug.Log("lane J touched");
-    }
-}
