--- Assets/Project/Scripts/PlayScreen/TouchEvent.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4038f135cc71c06428de5cc3fab31579
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
